dom/battery/test/test_battery_basics.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Battery API</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Battery API **/

ok('mozBattery' in navigator, "navigator.mozBattery should exist");

</script>
</pre>
</body>
</html>
